requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
drf-nested-routers==0.93.4